Move TrainingLogger to base.py
